+++ lib/pages/splash/splash.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
-class SplashPage extends StatefulWidget {
-  const SplashPage({super.key});
+class Splash extends StatefulWidget {
+  const Splash({super.key});
 
   @override
-  State<SplashPage> createState() => _SplashPageState();
+  State<Splash> createState() => _SplashState();
 }
 
-class _SplashPageState extends State<SplashPage> {
+class _SplashState extends State<Splash> {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
